remove goal panel
The goal panel revealed details about the reference implementation. Does it
employ REPEAT, IF, WHILE and recursion? This was supposed to give students
an estimate about the expected "language complexity" of the solution.

However, many students and supervisors interpreted the goal panel as a
binding contract. That is, if the reference implementation used REPEAT and
the student's solution didn't, the latter was deemed incorrect. This is
unfortunate, because I want students to come up with their own solutions
and not blindly target a set of somewhat arbitrary language features.

In the end, the goal panel did more harm than good, so I decided to pull it.